+++ lib/src/models/generator_options.g2.dart
@@ -11,8 +11,11 @@ GeneratorOptions _$GeneratorOptionsFromJson(Map json) => GeneratorOptions(
       withConverter: json['with_converter'] as bool? ?? true,
       ignoreHeaders: json['ignore_headers'] as bool? ?? false,
       separateModels: json['separate_models'] as bool? ?? false,
-      useDefaultNullForLists:
-          json['use_default_null_for_lists'] as bool? ?? false,
+      classesWithNullabeLists:
+          (json['classes_with_nullabe_lists'] as List<dynamic>?)
+                  ?.map((e) => e as String)
+                  .toList() ??
+              [],
       buildOnlyModels: json['build_only_models'] as bool? ?? false,
       defaultValuesMap: (json['default_values_map'] as List<dynamic>?)
               ?.map((e) =>
@@ -52,14 +55,32 @@ GeneratorOptions _$GeneratorOptionsFromJson(Map json) => GeneratorOptions(
               ?.map((e) => e as String)
               .toList() ??
           [],
+      nullableModels: (json['nullable_models'] as List<dynamic>?)
+              ?.map((e) => e as String)
+              .toList() ??
+          [],
+      cutFromModelNames: json['cut_from_model_names'] as String? ?? '',
+      additionalHeaders: (json['additional_headers'] as List<dynamic>?)
+              ?.map((e) => e as String)
+              .toList() ??
+          [],
+      overrideEqualsAndHashcode:
+          json['override_equals_and_hashcode'] as bool? ?? true,
+      overrideToString: json['override_to_string'] as bool? ?? true,
+      pageWidth: json['page_width'] as int?,
     );
 
 Map<String, dynamic> _$GeneratorOptionsToJson(GeneratorOptions instance) =>
     <String, dynamic>{
       'use_path_for_request_names': instance.usePathForRequestNames,
       'with_base_url': instance.withBaseUrl,
+      'page_width': instance.pageWidth,
+      'override_to_string': instance.overrideToString,
+      'override_equals_and_hashcode': instance.overrideEqualsAndHashcode,
       'with_converter': instance.withConverter,
+      'additional_headers': instance.additionalHeaders,
       'input_urls': instance.inputUrls,
+      'nullable_models': instance.nullableModels,
       'separate_models': instance.separateModels,
       'use_required_attribute_for_headers':
           instance.useRequiredAttributeForHeaders,
@@ -68,7 +89,8 @@ Map<String, dynamic> _$GeneratorOptionsToJson(GeneratorOptions instance) =>
       'include_if_null': instance.includeIfNull,
       'input_folder': instance.inputFolder,
       'output_folder': instance.outputFolder,
-      'use_default_null_for_lists': instance.useDefaultNullForLists,
+      'classes_with_nullabe_lists': instance.classesWithNullabeLists,
+      'cut_from_model_names': instance.cutFromModelNames,
       'build_only_models': instance.buildOnlyModels,
       'model_postfix': instance.modelPostfix,
       'default_values_map': instance.defaultValuesMap,